pyproject.toml
[project]
name = "plague-ingg"
version = "0.1.0"
description = "Plague InGG - A self-evolving AI agent with tools and memory"
authors = [
    {name = "Plague InGG", email = "plague-ingg@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "requests>=2.32.3",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pytest]
testpaths = ["tests"]

[tool.pylint]
max-line-length = 88

[tool.setuptools]
py-modules = ["llm_settings"]
